refactor!: move all methods to single file (`fr24.json` and `fr24.grpc`)
