+++ docs/api/functions.md
@@ -2,13 +2,9 @@ See the [example gallery](../usage/examples.md) to learn more.
 
 ::: fr24.authentication
 # JSON endpoints
-::: fr24.find
-::: fr24.history
+::: fr24.json
+
 # gRPC endpoints
-::: fr24.live_feed
-::: fr24.nearest_flights
-::: fr24.live_flights_status
-::: fr24.search_index
-::: fr24.follow_flight
-::: fr24.top_flights
-::: fr24.live_trail
+::: fr24.grpc
+    options:
+        show_if_no_docstring: true
