+++ docs/usage/scripts/27_search_index.py
@@ -4,9 +4,9 @@
 # %%
 # --8<-- [start:script0]
 import httpx
-from fr24.search_index import (
-    search_index_request_create,
-    search_index_post,
+from fr24.grpc import (
+    _search_index_request_create,
+    _search_index_post,
 )
 from fr24.proto.v1_pb2 import FetchSearchIndexRequest, FetchSearchIndexResponse
 
@@ -14,8 +14,8 @@ from fr24.proto.v1_pb2 import FetchSearchIndexRequest, FetchSearchIndexResponse
 async def search_index_data() -> FetchSearchIndexResponse:
     async with httpx.AsyncClient() as client:
         message = FetchSearchIndexRequest()
-        request = search_index_request_create(message)
-        return await search_index_post(client, request)
+        request = _search_index_request_create(message)
+        return await _search_index_post(client, request)
 
 
 data = await search_index_data()
