+++ src/fr24/grpc.py
@@ -1,22 +1,53 @@
+"""
+Endpoint: https://data-feed.flightradar24.com
+
+Service name: fr24.feed.api.v1.Feed
+
+Methods:
+
+- `LiveFeed`
+- `Playback`
+- `NearestFlights`
+- `LiveFlightsStatus`
+- `FollowFlight`
+- `TopFlights`
+- `LiveTrail`
+"""
+
 from __future__ import annotations
 
 import asyncio
-from typing import Any
+from typing import Any, AsyncIterator, Type
 
 import httpx
 from google.protobuf.field_mask_pb2 import FieldMask
 from loguru import logger
 
-from .proto import encode_message, parse_data
+from .proto import T, encode_message, parse_data
 from .proto.headers import get_headers
 from .proto.v1_pb2 import (
+    FetchSearchIndexRequest,
+    FetchSearchIndexResponse,
     Flight,
+    FollowFlightRequest,
+    FollowFlightResponse,
+    Geolocation,
+    HistoricTrailRequest,
+    HistoricTrailResponse,
     LiveFeedRequest,
     LiveFeedResponse,
+    LiveFlightsStatusRequest,
+    LiveFlightsStatusResponse,
+    LiveTrailRequest,
+    LiveTrailResponse,
     LocationBoundaries,
+    NearestFlightsRequest,
+    NearestFlightsResponse,
     PlaybackRequest,
     PlaybackResponse,
     RestrictionVisibility,
+    TopFlightsRequest,
+    TopFlightsResponse,
     TrafficType,
     VisibilitySettings,
 )
@@ -24,6 +55,41 @@ from .static.bbox import lng_bounds
 from .types.authentication import Authentication
 from .types.cache import LiveFeedRecord
 from .types.fr24 import LiveFeedField
+
+
+def construct_request(
+    method_name: str,
+    message: T,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    """Construct the gRPC request with encoded gRPC body."""
+    return httpx.Request(
+        "POST",
+        f"https://data-feed.flightradar24.com/fr24.feed.api.v1.Feed/{method_name}",
+        headers=get_headers(auth),
+        content=encode_message(message),
+    )
+
+
+async def post_unary(
+    client: httpx.AsyncClient, request: httpx.Request, msg_type: Type[T]
+) -> T:
+    """Execute the unary-unary call and return the parsed protobuf message."""
+    response = await client.send(request)
+    data = response.content
+    return parse_data(data, msg_type)
+
+
+async def post_stream(
+    client: httpx.AsyncClient, request: httpx.Request, msg_type: Type[T]
+) -> AsyncIterator[T]:
+    """Execute the unary-stream call and yield each parsed protobuf message."""
+    response = await client.send(request, stream=True)
+    try:
+        async for chunk in response.aiter_bytes():
+            yield parse_data(chunk, msg_type)
+    finally:
+        await response.aclose()
 
 
 def live_feed_message_create(
@@ -77,6 +143,19 @@ def live_feed_message_create(
     )
 
 
+def live_feed_request_create(
+    message: LiveFeedRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("LiveFeed", message, auth)
+
+
+async def live_feed_post(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> LiveFeedResponse:
+    return await post_unary(client, request, LiveFeedResponse)
+
+
 def live_feed_playback_message_create(
     message: LiveFeedRequest,
     timestamp: int,
@@ -98,48 +177,17 @@ def live_feed_playback_message_create(
     )
 
 
-def live_feed_request_create(
-    message: LiveFeedRequest,
-    auth: None | Authentication = None,
-) -> httpx.Request:
-    """Construct the POST request with encoded gRPC body."""
-    return httpx.Request(
-        "POST",
-        "https://data-feed.flightradar24.com/fr24.feed.api.v1.Feed/LiveFeed",
-        headers=get_headers(auth),
-        content=encode_message(message),
-    )
-
-
 def live_feed_playback_request_create(
     message: PlaybackRequest,
     auth: None | Authentication = None,
 ) -> httpx.Request:
-    """Constructs the POST request with encoded gRPC body."""
-    return httpx.Request(
-        "POST",
-        "https://data-feed.flightradar24.com/fr24.feed.api.v1.Feed/Playback",
-        headers=get_headers(auth),
-        content=encode_message(message),
-    )
-
-
-async def live_feed_post(
-    client: httpx.AsyncClient, request: httpx.Request
-) -> LiveFeedResponse:
-    """Send the request and parse the protobuf message."""
-    response = await client.send(request)
-    data = response.content
-    return parse_data(data, LiveFeedResponse)
+    return construct_request("Playback", message, auth)
 
 
 async def live_feed_playback_post(
     client: httpx.AsyncClient, request: httpx.Request
 ) -> PlaybackResponse:
-    """Send the request and parse raw protobuf message."""
-    response = await client.send(request)
-    data = response.content
-    return parse_data(data, PlaybackResponse)
+    return await post_unary(client, request, PlaybackResponse)
 
 
 def live_feed_flightdata_dict(
@@ -261,3 +309,116 @@ async def live_feed_playback_world_data(
         if not isinstance(r, BaseException)
         for lfr in r.live_feed_response.flights_list
     ]
+
+
+def nearest_flights_message_create(
+    lat: float,
+    lon: float,
+    radius: int,
+    limit: int = 1500,
+) -> NearestFlightsRequest:
+    """
+    Create the LiveFeedRequest protobuf message
+
+    :param lat: Centerpoint latitude (degrees)
+    :param lon: Centerpoint longitude (degrees)
+    :param radius: Circle radius (metres)
+    :param limit: Max number of flights
+    """
+    return NearestFlightsRequest(
+        location=Geolocation(lat=lat, lon=lon),
+        radius=radius,
+        limit=limit,
+    )
+
+
+def nearest_flights_request_create(
+    message: NearestFlightsRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("NearestFlights", message, auth)
+
+
+async def nearest_flights_post(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> NearestFlightsResponse:
+    return await post_unary(client, request, NearestFlightsResponse)
+
+
+def live_flights_status_request_create(
+    message: LiveFlightsStatusRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("LiveFlightsStatus", message, auth)
+
+
+async def live_flights_status_post(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> LiveFlightsStatusResponse:
+    return await post_unary(client, request, LiveFlightsStatusResponse)
+
+
+def _search_index_request_create(
+    message: FetchSearchIndexRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("FetchSearchIndex", message, auth)
+
+
+async def _search_index_post(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> FetchSearchIndexResponse:
+    return await post_unary(client, request, FetchSearchIndexResponse)
+
+
+def follow_flight_request_create(
+    message: FollowFlightRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("FollowFlight", message, auth)
+
+
+async def follow_flight_stream(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> AsyncIterator[FollowFlightResponse]:
+    async for msg in post_stream(client, request, FollowFlightResponse):
+        yield msg
+
+
+def top_flights_request_create(
+    message: TopFlightsRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("TopFlights", message, auth)
+
+
+async def top_flights_post(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> TopFlightsResponse:
+    return await post_unary(client, request, TopFlightsResponse)
+
+
+def live_trail_request_create(
+    message: LiveTrailRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("LiveTrail", message, auth)
+
+
+async def live_trail_post(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> LiveTrailResponse:
+    return await post_unary(client, request, LiveTrailResponse)
+
+
+def _historic_trail_request_create(
+    message: HistoricTrailRequest,
+    auth: None | Authentication = None,
+) -> httpx.Request:
+    return construct_request("HistoricTrail", message, auth)
+
+
+async def _historic_trail_post(
+    client: httpx.AsyncClient, request: httpx.Request
+) -> HistoricTrailResponse:
+    return await post_unary(client, request, HistoricTrailResponse)
