+++ src/fr24/proto/__init__.py
@@ -21,10 +21,9 @@ def encode_message(msg: T) -> bytes:
 # TODO: use Result<T, E> instead
 def parse_data(data: bytes, msg_type: Type[T]) -> T:
     assert len(data), "empty DATA frame"
-    assert data[0] == 0, "compressed message not implemented" # no compression
+    assert data[0] != 1, "compressed message not implemented" # no compression
+    assert data[0] == 0, f"received unknown message: {data}"
     data_len = int.from_bytes(data[1:5], byteorder="big") # length of message
     assert data_len > 0, "empty message"
     data_proto = data[5:5 + data_len]
-    # import base64
-    # print(data_len, base64.b64encode(data))
     return msg_type.FromString(data_proto)
