+++ src/fr24/tui/tui.py
@@ -21,12 +21,12 @@ from textual.widgets import (
 
 import pandas as pd
 from fr24.authentication import login
-from fr24.find import find, is_schedule
-from fr24.history import airport_list, flight_list, playback
+from fr24.json import airport_list, find, flight_list, playback
 from fr24.tui.formatters import Aircraft, Airport, Time
 from fr24.tui.widgets import AircraftWidget, AirportWidget, FlightWidget
 from fr24.types.airport_list import AirportList
 from fr24.types.authentication import Authentication
+from fr24.types.find import is_schedule
 from fr24.types.flight_list import FlightList, FlightListItem
 
 T = TypeVar("T")
