+++ tests/api/test_flight_list.py
@@ -7,7 +7,7 @@ import httpx
 import pyarrow.parquet as pq
 import pytest
 from fr24.core import FR24, FlightListArrow
-from fr24.history import flight_list, flight_list_df, playback
+from fr24.json import flight_list, flight_list_df, playback
 from fr24.types.flight_list import FlightList
 from pydantic import TypeAdapter
 
@@ -44,8 +44,8 @@ async def test_ll_flight_list() -> None:
 
 # core
 
-REG = "b-hpb"
-FLIGHT = "cx488"
+REG = "F-HEPK"
+FLIGHT = "AF7463"
 
 
 @pytest.mark.asyncio
